+++ src/mruby_integration/image.cpp
@@ -255,6 +255,16 @@ mrb_value mrb_image_colour_contrast(mrb_state *mrb, mrb_value) {
   return mrb_nil_value();
 }
 
+mrb_value mrb_image_colour_brightness(mrb_state *mrb, mrb_value) {
+  Image *image;
+  mrb_int brightness;
+  mrb_get_args(mrb, "di", &image, &Image_type, &brightness);
+
+  ImageColorBrightness(image, brightness);
+
+  return mrb_nil_value();
+}
+
 mrb_value mrb_image_resize_nearest_neighbour(mrb_state *mrb, mrb_value) {
   mrb_value image_obj;
   mrb_int width, height;
@@ -313,6 +323,7 @@ void append_images(mrb_state *mrb) {
   mrb_define_method(mrb, mrb->kernel_module, "image_colour_invert!", mrb_image_colour_invert, MRB_ARGS_REQ(1));
   mrb_define_method(mrb, mrb->kernel_module, "image_colour_grayscale!", mrb_image_colour_grayscale, MRB_ARGS_REQ(1));
   mrb_define_method(mrb, mrb->kernel_module, "image_colour_contrast!", mrb_image_colour_contrast, MRB_ARGS_REQ(2));
+  mrb_define_method(mrb, mrb->kernel_module, "image_colour_brightness!", mrb_image_colour_brightness, MRB_ARGS_REQ(2));
 
   mrb_define_method(mrb, mrb->kernel_module, "generate_image_colour", mrb_generate_image_colour, MRB_ARGS_REQ(3));